Commit da realização do atualizarConta e inativarConta, do Conta Service - 30/09/2025

+++ src/main/java/com/senai/conta_bancaria_spring_boot/Domain/Entity/ContaPoupanca.java
@@ -8,6 +8,8 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @EqualsAndHashCode(callSuper = true)
@@ -18,7 +20,7 @@ import lombok.experimental.SuperBuilder;
 public class ContaPoupanca extends Conta{
 
     @Column(precision = 5)
-    private double rendimento;
+    private BigDecimal rendimento;
 
     @Override
     public String getTipo(){
